ai-ml-ds/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
torch==2.1.0
transformers==4.35.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
pillow==10.1.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
